Feat: Removed dependency on SDL3_ttf due to its dependencies

+++ src/inc/ContentPanelSDL.h
@@ -31,7 +31,7 @@ public:
 	void tickOnClick();
 	void addComponent(int y, ComponentSDL* component);
 	void reposeContent();
-	void checkHoverOverContent(int& x, int& y);
+	void checkHoverOverContent(float& x, float& y);
 	void clear();
 };
 } /* namespace ConsoleArt */
